(fix) O3-4425 - make all top nav action items extensions
