+++ packages/apps/esm-implementer-tools-app/src/routes.json
@@ -14,7 +14,8 @@
     {
       "name": "implementer-tools-button",
       "slot": "top-nav-actions-slot",
-      "component": "implementerToolsButton"
+      "component": "implementerToolsButton",
+      "order": 10
     }
   ]
 }
