+++ packages/apps/esm-primary-navigation-app/src/components/navbar/navbar.component.tsx
@@ -1,28 +1,25 @@
-import React, { useCallback, useMemo, useState } from 'react';
-import { Navigate } from 'react-router-dom';
-import { useTranslation } from 'react-i18next';
-import classNames from 'classnames';
-import { HeaderContainer, Header, HeaderMenuButton, HeaderGlobalBar, HeaderGlobalAction } from '@carbon/react';
+import { Header, HeaderContainer, HeaderGlobalBar, HeaderMenuButton } from '@carbon/react';
 import {
-  useLayoutType,
-  ExtensionSlot,
   ConfigurableLink,
-  useSession,
+  ExtensionSlot,
   useAssignedExtensions,
   useConfig,
-  CloseIcon,
-  UserAvatarIcon,
-  SwitcherIcon,
+  useLayoutType,
   useLeftNavStore,
+  useSession,
 } from '@openmrs/esm-framework';
+import React, { useCallback, useMemo, useState } from 'react';
+import { useTranslation } from 'react-i18next';
+import { Navigate } from 'react-router-dom';
 import { isDesktop } from '../../utils';
-import AppMenuPanel from '../navbar-header-panels/app-menu-panel.component';
 import Logo from '../logo/logo.component';
 import NotificationsMenuPanel from '../navbar-header-panels/notifications-menu-panel.component';
-import OfflineBanner from '../offline-banner/offline-banner.component';
-import UserMenuPanel from '../navbar-header-panels/user-menu-panel.component';
 import SideMenuPanel from '../navbar-header-panels/side-menu-panel.component';
+import OfflineBanner from '../offline-banner/offline-banner.component';
+import AppMenuButton from './app-menu-button.component';
 import styles from './navbar.scss';
+import UserMenuButton from './user-menu-button.component';
+import { MenuButtonProps } from './navbar.resources';
 
 const HeaderItems: React.FC = () => {
   const { t } = useTranslation();
@@ -31,8 +28,6 @@ const HeaderItems: React.FC = () => {
   const layout = useLayoutType();
   const { slotName, mode } = useLeftNavStore();
   const navMenuItems = useAssignedExtensions(slotName);
-  const appMenuItems = useAssignedExtensions('app-menu-slot');
-  const userMenuItems = useAssignedExtensions('user-panel-slot');
   const isActivePanel = useCallback((panelName: string) => activeHeaderPanel === panelName, [activeHeaderPanel]);
 
   const togglePanel = useCallback((panelName: string) => {
@@ -50,8 +45,7 @@ const HeaderItems: React.FC = () => {
     () => (!isDesktop(layout) || mode === 'collapsed') && navMenuItems.length > 0,
     [navMenuItems.length, layout, mode],
   );
-  const showAppMenu = useMemo(() => appMenuItems.length > 0, [appMenuItems.length]);
-  const showUserMenu = useMemo(() => userMenuItems.length > 0, [userMenuItems.length]);
+
   return (
     <>
       <OfflineBanner />
@@ -75,7 +69,11 @@ const HeaderItems: React.FC = () => {
         <div className={styles.divider} />
         <ExtensionSlot name="top-nav-info-slot" />
         <HeaderGlobalBar className={styles.headerGlobalBar}>
-          <ExtensionSlot name="top-nav-actions-slot" className={styles.topNavActionsSlot} />
+          <ExtensionSlot
+            name="top-nav-actions-slot"
+            state={{ isActivePanel, togglePanel, hidePanel }}
+            className={styles.topNavActionsSlot}
+          />
           <ExtensionSlot
             name="notifications-menu-button-slot"
             state={{
@@ -83,47 +81,14 @@ const HeaderItems: React.FC = () => {
               togglePanel: togglePanel,
             }}
           />
-          {showUserMenu && (
-            <HeaderGlobalAction
-              aria-label={t('userMenuTooltip', 'My Account')}
-              aria-labelledby="Users Avatar Icon"
-              className={classNames({
-                [styles.headerGlobalBarButton]: isActivePanel('userMenu'),
-                [styles.activePanel]: !isActivePanel('userMenu'),
-              })}
-              enterDelayMs={500}
-              name="User"
-              isActive={isActivePanel('userMenu')}
-              onClick={() => {
-                togglePanel('userMenu');
-              }}
-            >
-              {isActivePanel('userMenu') ? <CloseIcon size={20} /> : <UserAvatarIcon size={20} />}
-            </HeaderGlobalAction>
-          )}
-          {showAppMenu && (
-            <HeaderGlobalAction
-              aria-label={t('AppMenuTooltip', 'App Menu')}
-              aria-labelledby="App Menu"
-              className={classNames({
-                [styles.headerGlobalBarButton]: isActivePanel('appMenu'),
-                [styles.activePanel]: !isActivePanel('appMenu'),
-              })}
-              enterDelayMs={500}
-              isActive={isActivePanel('appMenu')}
-              tooltipAlignment="end"
-              onClick={() => {
-                togglePanel('appMenu');
-              }}
-            >
-              {isActivePanel('appMenu') ? <CloseIcon size={20} /> : <SwitcherIcon size={20} />}
-            </HeaderGlobalAction>
-          )}
+          <ExtensionSlot
+            name="top-nav-app-menu-slot"
+            state={{ isActivePanel, togglePanel, hidePanel }}
+            className={styles.topNavActionsSlot}
+          />
+          <SideMenuPanel hidePanel={hidePanel('sideMenu')} expanded={isActivePanel('sideMenu')} />
+          <NotificationsMenuPanel expanded={isActivePanel('notificationsMenu')} />
         </HeaderGlobalBar>
-        <SideMenuPanel hidePanel={hidePanel('sideMenu')} expanded={isActivePanel('sideMenu')} />
-        {showAppMenu && <AppMenuPanel expanded={isActivePanel('appMenu')} hidePanel={hidePanel('appMenu')} />}
-        <NotificationsMenuPanel expanded={isActivePanel('notificationsMenu')} />
-        {showUserMenu && <UserMenuPanel expanded={isActivePanel('userMenu')} hidePanel={hidePanel('userMenu')} />}
       </Header>
     </>
   );
