+++ packages/apps/esm-primary-navigation-app/src/components/navbar/navbar.resources.ts
@@ -1,0 +1,5 @@
+export interface MenuButtonProps {
+  isActivePanel(panelName: string): boolean;
+  togglePanel(panelName: string);
+  hidePanel: (panelName: string) => () => void;
+}
