+++ packages/apps/esm-primary-navigation-app/src/root.component.test.tsx
@@ -47,20 +47,7 @@ jest.mock('./utils', () => ({
 describe('Root', () => {
   it('should display navbar with title', async () => {
     render(<Root />);
-
-    expect(screen.getByRole('button', { name: /My Account/i })).toBeInTheDocument();
-    expect(screen.getByRole('banner', { name: /openmrs/i })).toBeInTheDocument();
     expect(screen.getByText(/mock emr/i)).toBeInTheDocument();
-  });
-
-  it('should open user-menu panel', async () => {
-    const user = userEvent.setup();
-
-    render(<Root />);
-
-    const userButton = screen.getByRole('button', { name: /My Account/i });
-    await user.click(userButton);
-    expect(screen.getByLabelText(/location/i)).toBeInTheDocument();
   });
 
   describe('when view is desktop', () => {
